pyproject.toml
# ---- Project Info and Dependencies ----

[project.urls]
Repository = "https://github.com/rjpower/tidyllm"
# Homepage = "https://..."
# Documentation = "https://..."

[project]
name = "tidyllm"
description = "Easy LLM tool management."
authors = [
    { name="Russell Power", email="russell.power@gmail.com" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<4.0"
dynamic = ["version"]

# https://pypi.org/classifiers/
# Adjust as needed:
classifiers = [
    # Adjust as needed:
    "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    # Include this to avoid accidentally publishing to PyPI:
    "Private :: Do Not Upload",
]


# ---- Main dependencies ----

dependencies = [
    "click>=8.2.1",
    "fastapi>=0.115.14",
    "fastmcp>=2.10.1",
    "filetype>=1.2.0",
    "gdrive-fsspec>=0.2.0",
    "gdrivefs>=0.14.13",
    "genanki>=0.13.1",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "griffe>=1.7.3",
    "litellm>=1.73.6",
    "mcp[cli]>=1.10.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.2.1",
    "pypdfium2>=4.30.0",
    "pytest>=8.4.1",
    "pytest-xdist>=3.8.0",
    "questionary>=2.0.0",
    "rich>=14.0.0",
    "rpaframework>=30.0.2",
    "unidecode>=1.4.0",
]


# ---- Dev dependencies ----

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-sugar>=1.0.0",
    "ruff>=0.11.9",
    "rich>=14.0.0",
    "funlog>=0.2.1",
]

[project.scripts]
tidyllm = "scripts.run_tidyllm:main"
tidyapi = "scripts.run_fastapi:main"
tidymcp = "scripts.run_fastmcp:main"

[project.optional-dependencies]
audio = [
    "pydub>=0.25.1",
    "torch>=2.0.0",
    "silero-vad>=5.0.0",
    "librosa>=0.10.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
]

automation = [
    "rpaframework>=29.0.0",
]

browser = [
    "playwright>=1.40.0",
]


# ---- Build system ----

# Dynamic versioning from:
# https://github.com/ninoseki/uv-dynamic-versioning/

[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"
# Note JSON schemas don't seem to be right for tool.hatch.version.source so
# this may cause false warnings in IDEs.
# https://github.com/ninoseki/uv-dynamic-versioning/issues/21

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = "true"

[tool.hatch.build.targets.wheel]
packages = ["src/tidyllm", "scripts"]


# ---- Settings ----

[tool.ruff]
# Set as desired, typically 88 (black standard) or 100 (wide).
line-length = 100
exclude = ["tests/"]


[tool.ruff.lint]
select = [
    # See: https://docs.astral.sh/ruff/rules/
    # Basic list from: https://docs.astral.sh/ruff/linter/#rule-selection
    "E", # https://docs.astral.sh/ruff/rules/#error-e
    "F", # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "UP", # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "B", # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "I", # https://docs.astral.sh/ruff/rules/#isort-i
    # Other possibilities:
    # "D" # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    # "Q" # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    # "COM" # https://docs.astral.sh/ruff/rules/#flake8-commas-com
    # "SIM", # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim

]
ignore = [
    # Disable some rules that are overly pedantic. Add/remove as desired:
    "E501", # https://docs.astral.sh/ruff/rules/line-too-long/
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    # We use both ruff formatter and linter so some rules should always be disabled.
    # See: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191", # https://docs.astral.sh/ruff/rules/tab-indentation/
    "E111", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple/
    "E114", # https://docs.astral.sh/ruff/rules/indentation-with-invalid-multiple-comment/
    "E117", # https://docs.astral.sh/ruff/rules/over-indented/
    "D206", # https://docs.astral.sh/ruff/rules/docstring-tab-indentation/
    "D300", # https://docs.astral.sh/ruff/rules/triple-single-quotes/
    "Q000", # https://docs.astral.sh/ruff/rules/bad-quotes-inline-string/
    "Q001", # https://docs.astral.sh/ruff/rules/bad-quotes-multiline-string/
    "Q002", # https://docs.astral.sh/ruff/rules/bad-quotes-docstring/
    "Q003", # https://docs.astral.sh/ruff/rules/avoidable-escaped-quote/
    "COM812", # https://docs.astral.sh/ruff/rules/missing-trailing-comma/
    "COM819", # https://docs.astral.sh/ruff/rules/prohibited-trailing-comma/
    "ISC002", # https://docs.astral.sh/ruff/rules/multi-line-implicit-string-concatenation/
]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = ["**/*venv/**/*", "tests/**/*"]


[tool.pytest.ini_options]
python_files = ["*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = [
    "tests",
]
norecursedirs = []
markers = [
    "integration: marks for integration testing"
]
filterwarnings = [
   "ignore::DeprecationWarning:aifc",
   "ignore::DeprecationWarning:audioop", 
   "ignore::DeprecationWarning:sunau",
   "ignore::DeprecationWarning:audioread.*",
   "ignore::DeprecationWarning:_pytest.*",
]
addopts = [ "-n", "auto" ]

[tool.uv.sources]
gdrive-fsspec = { git = "https://github.com/fsspec/gdrivefs" }
